@@ -1,7 +1,8 @@
 package com.main.suwoninfo.domain;
 import lombok.*;
 
-import javax.persistence.*;
+import jakarta.persistence.*;
+
 
 @Entity
 @Getter
