@@ -1,36 +1,36 @@
 package com.main.suwoninfo.jwt;
 
-import com.fasterxml.jackson.databind.ObjectMapper;
-import com.main.suwoninfo.exception.CustomException;
 import com.main.suwoninfo.form.TokenResponse;
-import com.main.suwoninfo.service.UserService;
 import com.main.suwoninfo.utils.RedisUtils;
+import jakarta.servlet.FilterChain;
+import jakarta.servlet.ServletException;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletRequestWrapper;
+import jakarta.servlet.http.HttpServletResponse;
 import lombok.RequiredArgsConstructor;
+import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.util.StringUtils;
-import org.springframework.web.filter.GenericFilterBean;
+import org.springframework.web.client.RestClient;
 import org.springframework.web.filter.OncePerRequestFilter;
 
-import javax.servlet.FilterChain;
-import javax.servlet.ServletException;
-import javax.servlet.ServletRequest;
-import javax.servlet.ServletResponse;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
+import java.util.Map;
 
 @RequiredArgsConstructor
 public class JwtFilter extends OncePerRequestFilter {
 
     private final JwtTokenProvider tokenProvider;
-    private final UserService userService;
     private final RedisUtils redisUtils;
+    private final RestClient restClient;
 
     @Override
     protected void doFilterInternal(HttpServletRequest request, HttpServletResponse response, FilterChain filterChain)
             throws ServletException, IOException {
-        String token = resolveToken(request);
+
+        HttpServletRequestWrapper requestWrapper = new CachingBodyRequestWrapper(request);
+        String token = resolveToken(requestWrapper);
 
         try {
             if (token != null) {
@@ -41,8 +41,8 @@ public class JwtFilter extends OncePerRequestFilter {
                     System.out.println("Authentication set: " + authentication.getName());
                 } else if (tokenProvider.isTokenExpired(token)) {
                     // Access Token 만료 시 처리
-                    Boolean reissued = handleExpiredToken(request, response, token);
-                    if (!reissued) return;
+                    HttpServletRequest reissued = handleExpiredToken(requestWrapper, response, token);
+                    if (reissued != null) filterChain.doFilter(reissued, response);
                 } else {
                     response.sendError(HttpServletResponse.SC_UNAUTHORIZED, "INVALID_TOKEN");
                     return;
@@ -58,13 +58,13 @@ public class JwtFilter extends OncePerRequestFilter {
         }
     }
 
-    private Boolean handleExpiredToken(HttpServletRequest request, HttpServletResponse response, String expiredToken)
+    private HttpServletRequest handleExpiredToken(HttpServletRequestWrapper request, HttpServletResponse response, String expiredToken)
             throws IOException {
         Authentication authentication = tokenProvider.getAuthentication(expiredToken);
 
         if (authentication == null) {
             response.sendError(HttpServletResponse.SC_UNAUTHORIZED, "INVALID_TOKEN");
-            return false;
+            return null;
         }
 
         String email = authentication.getName();
@@ -74,35 +74,34 @@ public class JwtFilter extends OncePerRequestFilter {
             refreshToken = refreshToken.replace("\"", "");
             if (!tokenProvider.validateToken(refreshToken).getValid()) {
                 response.sendError(HttpServletResponse.SC_UNAUTHORIZED, "INVALID_TOKEN");
-                return false;
+                return null;
             }
         } else {
             response.sendError(HttpServletResponse.SC_UNAUTHORIZED, "INVALID_TOKEN");
-            return false;
+            return null;
         }
 
+        TokenResponse tokenResponse = TokenResponse.builder()
+                .accessToken(expiredToken)
+                .refreshToken(refreshToken)
+                .tokenType("Bearer")
+                .build();
+
         try {
-            // 새로운 Access Token 발급
-            TokenResponse newTokens = userService.reissue(expiredToken, refreshToken);
-
-            response.setHeader("Authorization", "Bearer " + newTokens.getAccessToken());
-            redisUtils.set("RT:" + email, newTokens.getRefreshToken(), 4320); // Redis에 Refresh Token 갱신
-
-            response.setContentType("application/json");
-            response.setCharacterEncoding("UTF-8");
-
-            String json = new ObjectMapper().writeValueAsString(newTokens);
-            response.getWriter().write(json);
-
-            // SecurityContextHolder에 새로운 인증 설정
-            Authentication newAuthentication = tokenProvider.getAuthentication(newTokens.getAccessToken());
-            SecurityContextHolder.getContext().setAuthentication(newAuthentication);
-            return true;
+            ResponseEntity<TokenResponse> tokenEntity = restClient
+                    .post()
+                    .uri("http://localhost:8081/users/reissue")
+                    .header("Authorization", "Bearer " + expiredToken)
+                    .body(tokenResponse)
+                    .retrieve()
+                    .toEntity(TokenResponse.class);
+            Map <String, String> mutatedRequestHeader = Map.of("Authorization", "Bearer " + tokenEntity.getBody().getAccessToken());
+            return new CachingBodyRequestWrapper(request, mutatedRequestHeader);
 
         } catch (Exception e) {
             response.sendError(HttpServletResponse.SC_UNAUTHORIZED, "token generation failed");
             System.out.println("에러" + e.getMessage());
-            return false;
+            return null;
         }
     }
 
