@@ -2,10 +2,9 @@ package com.main.suwoninfo.repository;
 
 import com.main.suwoninfo.domain.Comment;
 import com.querydsl.jpa.impl.JPAQueryFactory;
+import jakarta.persistence.EntityManager;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Repository;
-
-import javax.persistence.EntityManager;
 
 import java.util.List;
 import java.util.Optional;
