@@ -2,13 +2,10 @@ package com.main.suwoninfo.repository;
 
 import com.main.suwoninfo.domain.User;
 import com.main.suwoninfo.domain.UserAuthority;
-import com.main.suwoninfo.exception.CustomException;
-import com.main.suwoninfo.exception.UserErrorCode;
-import com.querydsl.jpa.impl.JPAQueryFactory;
+import jakarta.persistence.EntityManager;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Repository;
 
-import javax.persistence.EntityManager;
 import java.util.List;
 import java.util.Optional;
 
@@ -18,7 +15,6 @@ import java.util.Optional;
 public class UserRepository {
 
     private final EntityManager entityManager;
-    private final JPAQueryFactory queryFactory;
 
 
     public void save(User user) {
